feat: merge the tablebodycell component into tablecell for simplier usage (#1219)
* feat: merge the tablebodycell component into tablecell for simplier usage

* feat: Improved logic for the expendable column

* fix: changed the hardcoded way of styling by using scss for styling

* fix: fix lint error in scss file
