+++ src/components/DataTable/cells/TableCell.tsx
@@ -1,32 +1,79 @@
 import React from 'react';
 import clsx from 'clsx';
+import {IconTextIcon, Typography} from '~/components';
+import {ChevronDown, ChevronRight} from '~/icons';
 
 import type {TableCellProps} from './TableCell.types';
 import {capitalize} from '~/utils/helpers';
 import './TableCell.scss';
 
-/**
- * Base table cell component.
- *
- * Usage 1: Direct usage with children prop → renders children directly.
- */
 const TableCellForwardRef: React.ForwardRefRenderFunction<HTMLTableCellElement, TableCellProps> = (
-    {className, children, textAlign, ...props},
+    {
+        className,
+        children,
+        textAlign = 'left',
+        verticalAlign = 'center',
+        iconStart,
+        iconEnd,
+        isExpandableColumn,
+        row,
+        width,
+        isScrollable,
+        component = 'td',
+        ...props
+    },
     ref
 ) => {
+    const leftMarginBuffer = 20; // Px
+    const leftMarginIndentDepth = (row?.depth ?? 0) * 20; // Px
+    const scrollableClass = isScrollable ? 'moonstone-tableCellContent' : '';
+
+    const renderCellContent = () => (
+        <IconTextIcon component="div" iconStart={iconStart} iconEnd={iconEnd} typographyProps={{className: clsx(scrollableClass, 'flexFluid')}}>
+            {children ?? '-'}
+        </IconTextIcon>
+    );
+
+    const renderExpandableContent = () => {
+        if (row?.getCanExpand?.()) {
+            return (
+                <div
+                    className="moonstone-tableCellExpandable flexRow_nowrap alignCenter"
+                    style={{marginLeft: `${leftMarginIndentDepth}px`}}
+                    onClick={row.getToggleExpandedHandler()}
+                >
+                    {row.getIsExpanded() ?
+                        <ChevronDown className="moonstone-marginRightNano"/> :
+                        <ChevronRight className="moonstone-marginRightNano"/>}
+                    {renderCellContent()}
+                </div>
+            );
+        }
+
+        return (
+            <div style={{marginLeft: `${leftMarginIndentDepth + leftMarginBuffer}px`}}>
+                {renderCellContent()}
+            </div>
+        );
+    };
+
     return (
-        <td
+        <Typography
             ref={ref}
+            component={component}
+            variant="body"
             className={clsx(
                 'moonstone-TableCell',
                 'textAlign' + capitalize(textAlign),
+                'moonstone-verticalAlign' + capitalize(verticalAlign),
+                {flexFluid: typeof width === 'undefined'},
                 className
             )}
+            style={{width: width}}
             {...props}
         >
-            {/* If children provided, use it; otherwise show "-" */}
-            {children ?? '-'}
-        </td>
+            {isExpandableColumn ? renderExpandableContent() : renderCellContent()}
+        </Typography>
     );
 };
