+++ src/components/DataTable/table-cells/index.ts
@@ -1,2 +1,1 @@
-export * from './TableBodyCell';
 export * from './TableHeadCell';
